+++ R/modifySoilParams.R
@@ -2,9 +2,9 @@
 #'
 #' @param soildata A data frame of soil physical parameters or a list of data frames
 #' @param soildepth An estimate of soil depth (in mm) or a vector of soil depths if `soildata` is a list
-#' @param surfacerock An estimate of surface rock content (in percent) or a vector of such values if `soildata` is a list
 #'
 #' @return A data frame of modified physical parameters or a list of modified data frames
+#' @name modifySoilParams
 #' @export
 #'
 #' @examples
@@ -13,17 +13,16 @@
 #'   library(sp)
 #'   coords_sp <- SpatialPoints(cbind(long = -5.6333, lat = 42.6667),
 #'                                  CRS(SRS_string = "EPSG:4326"))
-#'   foo <- soilgridsParams(coords_sp, widths = c(300, 700, 1000, 2000))
-#'   foo
-#'   foo_mod <- modifySoilParams(foo, 900, 20) # 90 cm depth, 20 % rocks in the surface
-#'   foo_mod
+#'   foo_1 <- soilgridsParams(coords_sp, widths = c(300, 700, 1000, 2000))
+#'   foo_1
+#'   foo_2 <- modifySoilDepth(foo_1, 900) # 90 cm depth
+#'   foo_2
+#'   foo_3 <- modifySoilRockContent(foo_2, 20) # 20 % rocks in the surface
+#'   foo_3
 #'  }
-modifySoilParams<-function(soildata, soildepth, surfacerock = NULL) {
-  rock_from_surface<-function(surfacerock, depth) {
-    coef = 1+3^(depth/1000)
-    return(100*(1 - exp(-(max(10,surfacerock)/100)*coef)))
-  }
-  modifySoilOne <-function(soildf, soildepthone, surfacerockone) {
+#'
+modifySoilDepth<-function(soildata, soildepth) {
+  modifySoilDepthOne <-function(soildf, soildepthone) {
     soildepthone = pmax(50,soildepthone)
     nl = nrow(soildf)
     oridepths = c(0, cumsum(soildf$widths[1:(nl-1)]))
@@ -42,29 +41,54 @@ modifySoilParams<-function(soildata, soildepth, surfacerock = NULL) {
       nl = nrow(soildf)
       soildf$widths[nl] = orimaxdepth - soildepthone
     }
-    if(!is.null(surfacerockone)) {
-      for(l in 1:(nl-1)) {
-        upper = 0
-        if(l>1) upper = sum(soildf$widths[1:(l-1)])
-        lower = sum(soildf$widths[1:l])
-        middepth = 0.5*(upper+lower)
-        soildf$rfc[l] = min(97.5, max(soildf$rfc[l], rock_from_surface(surfacerockone, middepth)))
-      }
-    }
     soildf$rfc[nl] = 97.5
     return(soildf)
   }
   if(inherits(soildata, "data.frame")) {
     if(length(soildepth)>1) stop("'soildepth' has to be of length one when 'soildata' is a data.frame")
-    if(!is.null(surfacerock)) if(length(surfacerock)>1) stop("'surfacerock' has to be of length one when 'soildata' is a data.frame")
-    return(modifySoilOne(soildata, soildepth, surfacerock))
+    return(modifySoilDepthOne(soildata, soildepth))
   } else if(inherits(soildata, "list")) {
     if(length(soildepth)!=length(soildata)) stop("Vector 'soildepth' has to be of the same length as 'soildata'")
+    for(i in 1:length(soildata)) {
+      if(inherits(soildata[[i]], "data.frame")) {
+        soildata[[i]] = modifySoilDepthOne(soildata[[i]], soildepth[i])
+      } else {
+        stop("Wrong class for soil data. Has to be a data.frame or a list of data.frame objects")
+      }
+    }
+    return(soildata)
+  } else {
+    stop("Wrong class for soil data. Has to be a data.frame or a list of data.frame objects")
+  }
+}
+
+#' @param surfacerock An estimate of surface rock content (in percent) or a vector of such values if `soildata` is a list
+#'
+#' @rdname modifySoilParams
+modifySoilRockContent<-function(soildata, surfacerock) {
+  rock_from_surface<-function(surfacerock, depth) {
+    coef = 1+3^(depth/1000)
+    return(100*(1 - exp(-(max(10,surfacerock)/100)*coef)))
+  }
+  modifySoilRockContentOne <-function(soildf, surfacerockone) {
+    nl = nrow(soildf)
+    for(l in 1:(nl-1)) {
+      upper = 0
+      if(l>1) upper = sum(soildf$widths[1:(l-1)])
+      lower = sum(soildf$widths[1:l])
+      middepth = 0.5*(upper+lower)
+      soildf$rfc[l] = min(97.5, max(soildf$rfc[l], rock_from_surface(surfacerockone, middepth)))
+    }
+    return(soildf)
+  }
+  if(inherits(soildata, "data.frame")) {
+    if(!is.null(surfacerock)) if(length(surfacerock)>1) stop("'surfacerock' has to be of length one when 'soildata' is a data.frame")
+    return(modifySoilRockContentOne(soildata, surfacerock))
+  } else if(inherits(soildata, "list")) {
     if(!is.null(surfacerock)) if(length(surfacerock)!=length(soildata)) stop("Vector 'surfacerock' has to be of the same length as 'soildata'")
     for(i in 1:length(soildata)) {
       if(inherits(soildata[[i]], "data.frame")) {
-        if(!is.null(surfacerock)) soildata[[i]] = modifySoilOne(soildata[[i]], soildepth[i], surfacerock[i])
-        else soildata[[i]] = modifySoilOne(soildata[[i]], soildepth[i], NULL)
+        soildata[[i]] = modifySoilRockContentOne(soildata[[i]], surfacerock[i])
       } else {
         stop("Wrong class for soil data. Has to be a data.frame or a list of data.frame objects")
       }