+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 0.1.4
 - Names changed for IFN example datasets
 - Update of function 'forest_mapShrubTable'
+- Function split 'modifySoilDepth' and 'modifySoilRockContent'
 
 # Version 0.1.3
 - Keep num orden from IFN data